+++ src/main/java/com/ats/webapi/model/Orders.java
@@ -69,7 +69,7 @@ public class Orders {
 	private int userId;
 	
 	@Column(name="is_positive")
-	private int isPositive;
+	private float isPositive;
 
 	@Column(name="menu_id")
 	private int menuId;
@@ -235,11 +235,11 @@ public class Orders {
 		this.userId = userId;
 	}
 
-	public int getIsPositive() {
+	public float getIsPositive() {
 		return isPositive;
 	}
 
-	public void setIsPositive(int isPositive) {
+	public void setIsPositive(float isPositive) {
 		this.isPositive = isPositive;
 	}
